feat: show language statistics

+++ app/app_statistics/models.py
@@ -11,9 +11,9 @@ class Statistics(models.Model):
     audios_approved = models.IntegerField(default=0)
     audios_transcribed = models.IntegerField(default=0)
 
-    audios_hours_submitted = models.IntegerField(default=0)
-    audios_hours_approved = models.IntegerField(default=0)
-    audios_hours_transcribed = models.IntegerField(default=0)
+    audios_hours_submitted = models.DecimalField(default=0, decimal_places=2, max_digits=20)
+    audios_hours_approved = models.DecimalField(default=0, decimal_places=2, max_digits=20)
+    audios_hours_transcribed = models.DecimalField(default=0, decimal_places=2, max_digits=20)
 
     ########### EWE ########################
     # Total Count
@@ -25,12 +25,12 @@ class Statistics(models.Model):
     ewe_audios_transcribed = models.IntegerField(default=0)
 
     # Total Hours
-    ewe_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ewe_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ewe_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ewe_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ewe_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ewe_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
+    ewe_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ewe_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ewe_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ewe_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ewe_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ewe_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
 
 
     ########### AKAN ########################
@@ -43,12 +43,12 @@ class Statistics(models.Model):
     akan_audios_transcribed = models.IntegerField(default=0)
 
     # Total Hours
-    akan_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    akan_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    akan_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    akan_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    akan_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    akan_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
+    akan_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    akan_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    akan_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    akan_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    akan_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    akan_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
 
 
     ########### DAGBANI ########################
@@ -61,12 +61,12 @@ class Statistics(models.Model):
     dagbani_audios_transcribed = models.IntegerField(default=0)
 
     # Total Hours
-    dagbani_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagbani_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagbani_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagbani_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagbani_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagbani_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
+    dagbani_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagbani_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagbani_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagbani_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagbani_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagbani_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
 
     ########### DAGAARE ########################
     # Total Count
@@ -78,12 +78,12 @@ class Statistics(models.Model):
     dagaare_audios_transcribed = models.IntegerField(default=0)
 
     # Total Hours
-    dagaare_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagaare_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagaare_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagaare_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagaare_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    dagaare_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
+    dagaare_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagaare_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagaare_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagaare_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagaare_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    dagaare_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
 
     ########### IKPOSO ########################
     # Total Count
@@ -95,12 +95,12 @@ class Statistics(models.Model):
     ikposo_audios_transcribed = models.IntegerField(default=0)
 
     # Total Hours
-    ikposo_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ikposo_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ikposo_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ikposo_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ikposo_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
-    ikposo_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=4, max_digits=20)
+    ikposo_audios_submitted_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ikposo_audios_single_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ikposo_audios_double_validation_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ikposo_audios_validation_conflict_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ikposo_audios_approved_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
+    ikposo_audios_transcribed_in_hours = models.DecimalField(default=0.0, decimal_places=2, max_digits=20)
 
     # Timestamp
     created_at = models.DateTimeField(auto_now_add=True)
